Remove unneeded in-between component
